Split phpliteadmin.php in multiple files - Fixed Issue #190 - split classes in separate files, under classes/ - moved language files under languages/ - split project/license info in docs/header.txt - added build script and template - added support code to minify js and css, under support/ - *** phpliteadmin.php has been renamed index.php ***
As such, the repository does not contain a file named phpliteadmin.php.

The index.php script is not standalone, but requires the following files
to run:
- phpliteadmin.config.sample.php
- language/lang_en.php
- classes/*
- resources/*

A single-file version of phpLiteAdmin, equivalent to previous releases,
can be obtained by running build.php. The built script will be saved as
phpliteadmin.php and requires no other files to run, just as before.



git-svn-id: http://phpliteadmin.googlecode.com/svn/source/1.9.5@385 c054f44c-0d31-6418-e286-147a9414beb1